move the sprite vocabulary and pure track modules below validation
lib/spriteValidation.js imported six sprite service modules to enumerate the
alphabets its Zod schemas validate against. Same call as the FableLoom/POST
slice — push the dependency down — because promoting the schemas was not an
option: lib/validation.js re-exports this module, so moving it up would have
inverted that instead.

Three modules move wholesale, each already a leaf by its own description:

  services/sprites/chromaKey.js          -> lib/spriteChromaKey.js
  services/sprites/animationTracks.js    -> lib/spriteAnimationTracks.js
  services/sprites/animationTrackStore.js-> lib/spriteAnimationTrackStore.js

chromaKey's own header called it "pure … safe to import anywhere"; the track
registry imports nothing at all; the store reads one small JSON config through
lib/fileUtils and depends on no service. The four remaining scattered constants
(sprite id pattern, record kinds, the 8-direction order and its derived anchor
set, TURNAROUND_ID, the animation provider ids) collect into the new
lib/spriteVocabulary.js. Old paths re-export throughout.

QUEUEABLE_IMAGE_MODES now comes from lib/generationModes.js, where #5021 put it.

Three test consequences the move surfaced, all fixed here rather than papered
over:

- The sharp-free property test asserts the request-validation graph never
  reaches a native image dependency. It derived its roots from its own file
  location, so moving it broke the path model; both roots are now named
  explicitly and each entry points at the module's real home. The guard still
  walks the same graph.
- Its vi.mock of lib/fileUtils.js stopped resolving after the move, so the store
  test read (and wrote) the real data dir instead of its fixture. Repointed.
- animationTargets.test.js mocked services/sprites/animationTracks.js to stand
  up a null-fps row, but the row it resolves comes from the store, which now
  imports the lib registry directly — so the mock no longer reached it and the
  test had silently stopped discriminating. Repointed at the real seam.

Refs #4901.
