+++ server/lib/index.js
@@ -349,6 +349,10 @@ export * from './markdownText.js';
 export * from './renderSlot.js';
 export * from './renderTargets.js';
 export * from './generationModes.js';
+export * from './spriteVocabulary.js';
+export * from './spriteChromaKey.js';
+export * from './spriteAnimationTracks.js';
+export * from './spriteAnimationTrackStore.js';
 export * from './postDrillTypes.js';
 export * from './telegramClient.js';
 export * from './tempPathGuard.js';
